+++ datapipe/compute.py
@@ -5,10 +5,11 @@ from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import Dict, Iterable, List, Optional, Tuple
 
-import tqdm
 from opentelemetry import trace
+from tqdm_loggable.auto import tqdm
 
 from datapipe.datatable import DataStore, DataTable
+from datapipe.executor import Executor, ExecutorConfig, SingleThreadExecutor
 from datapipe.run_config import RunConfig
 from datapipe.store.table_store import TableStore
 from datapipe.types import ChangeList, DataDF, IndexDF, Labels, TransformResult
@@ -63,11 +64,13 @@ class ComputeStep:
         input_dts: List[DataTable],
         output_dts: List[DataTable],
         labels: Optional[Labels] = None,
+        executor_config: Optional[ExecutorConfig] = None,
     ) -> None:
         self._name = name
         self.input_dts = input_dts
         self.output_dts = output_dts
         self._labels = labels
+        self.executor_config = executor_config
 
     def get_name(self) -> str:
         ss = [
@@ -122,6 +125,7 @@ class ComputeStep:
     def get_full_process_ids(
         self,
         ds: DataStore,
+        chunk_size: Optional[int] = None,
         run_config: Optional[RunConfig] = None,
     ) -> Tuple[int, Iterable[IndexDF]]:
         raise NotImplementedError()
@@ -258,23 +262,29 @@ class ComputeStep:
         self,
         ds: DataStore,
         run_config: Optional[RunConfig] = None,
+        executor: Optional[Executor] = None,
     ) -> None:
+        if executor is None:
+            executor = SingleThreadExecutor()
+
         logger.info(f"Running: {self.name}")
         run_config = RunConfig.add_labels(run_config, {"step_name": self.name})
 
-        (idx_count, idx_gen) = self.get_full_process_ids(ds, run_config)
+        (idx_count, idx_gen) = self.get_full_process_ids(ds=ds, run_config=run_config)
 
         logger.info(f"Batches to process {idx_count}")
 
         if idx_count is not None and idx_count == 0:
             return
 
-        for idx in tqdm.tqdm(idx_gen, total=idx_count):
-            self.process_batch(
-                ds=ds,
-                idx=idx,
-                run_config=run_config,
-            )
+        executor.run_process_batch(
+            ds=ds,
+            idx_count=idx_count,
+            idx_gen=idx_gen,
+            process_fn=self.process_batch,
+            run_config=run_config,
+            executor_config=self.executor_config,
+        )
 
         ds.event_logger.log_step_full_complete(self.name)
 
@@ -283,7 +293,11 @@ class ComputeStep:
         ds: DataStore,
         change_list: ChangeList,
         run_config: Optional[RunConfig] = None,
+        executor: Optional[Executor] = None,
     ) -> ChangeList:
+        if executor is None:
+            executor = SingleThreadExecutor()
+
         logger.info(f"Running: {self.name}")
         run_config = RunConfig.add_labels(run_config, {"step_name": self.name})
 
@@ -298,7 +312,7 @@ class ComputeStep:
 
         res_changelist = ChangeList()
 
-        for idx in tqdm.tqdm(idx_gen, total=idx_count):
+        for idx in tqdm(idx_gen, total=idx_count):
             changes = self.process_batch(
                 ds=ds,
                 idx=idx,
@@ -307,6 +321,21 @@ class ComputeStep:
             res_changelist.extend(changes)
 
         return res_changelist
+
+    def run_idx(
+        self,
+        ds: DataStore,
+        idx: IndexDF,
+        run_config: Optional[RunConfig] = None,
+    ) -> ChangeList:
+        logger.info(f"Running: {self.name}")
+        run_config = RunConfig.add_labels(run_config, {"step_name": self.name})
+
+        return self.process_batch(
+            ds=ds,
+            idx=idx,
+            run_config=run_config,
+        )
 
 
 class PipelineStep(ABC):
@@ -361,7 +390,10 @@ def print_compute(steps: List[ComputeStep]) -> None:
 
 
 def run_steps(
-    ds: DataStore, steps: List[ComputeStep], run_config: Optional[RunConfig] = None
+    ds: DataStore,
+    steps: List[ComputeStep],
+    run_config: Optional[RunConfig] = None,
+    executor: Optional[Executor] = None,
 ) -> None:
     for step in steps:
         with tracer.start_as_current_span(
@@ -372,7 +404,7 @@ def run_steps(
                 f"{[i.name for i in step.input_dts]} -> {[i.name for i in step.output_dts]}"
             )
 
-        step.run_full(ds, run_config)
+        step.run_full(ds=ds, run_config=run_config, executor=executor)
 
 
 def run_pipeline(