+++ datapipe/metastore.py
@@ -7,7 +7,7 @@ from typing import Iterator, List, Optional, Tuple, cast
 
 import pandas as pd
 from cityhash import CityHash32
-from sqlalchemy import Boolean, Column, Float, Integer, String, Table, func
+from sqlalchemy import Boolean, Column, Float, Integer, String, Table, func, update
 from sqlalchemy.sql.expression import and_, delete, or_, select, text, tuple_
 
 from datapipe.run_config import RunConfig
@@ -138,21 +138,23 @@ class MetaTable:
         res = []
         sql = select(self.sql_schema)
 
-        if idx is None:
-            sql = self._build_metadata_query(sql, idx, include_deleted)
-            return cast(MetadataDF, pd.read_sql_query(sql, con=self.dbconn.con))
+        with self.dbconn.con.begin() as con:
+            if idx is None:
+                sql = self._build_metadata_query(sql, idx, include_deleted)
+                return cast(MetadataDF, pd.read_sql_query(sql, con=con))
 
-        for chunk_idx in self._chunk_idx_df(idx):
-            chunk_sql = self._build_metadata_query(sql, chunk_idx, include_deleted)
-            res.append(pd.read_sql_query(chunk_sql, con=self.dbconn.con))
+            for chunk_idx in self._chunk_idx_df(idx):
+                chunk_sql = self._build_metadata_query(sql, chunk_idx, include_deleted)
 
-        if len(res) > 0:
-            return cast(MetadataDF, pd.concat(res))
-        else:
-            return cast(
-                MetadataDF,
-                pd.DataFrame(columns=[column.name for column in self.sql_schema]),
-            )
+                res.append(pd.read_sql_query(chunk_sql, con=con))
+
+            if len(res) > 0:
+                return cast(MetadataDF, pd.concat(res))
+            else:
+                return cast(
+                    MetadataDF,
+                    pd.DataFrame(columns=[column.name for column in self.sql_schema]),
+                )
 
     def get_metadata_size(
         self, idx: Optional[IndexDF] = None, include_deleted: bool = False
@@ -167,8 +169,9 @@ class MetaTable:
         sql = select([func.count()]).select_from(self.sql_table)
         sql = self._build_metadata_query(sql, idx, include_deleted)
 
-        res = self.dbconn.con.execute(sql).fetchone()
-        return res[0]
+        with self.dbconn.con.begin() as con:
+            res = con.execute(sql).fetchone()
+            return res[0]
 
     def _make_new_metadata_df(self, now: float, df: DataDF) -> MetadataDF:
         meta_keys = self.primary_keys + list(self.meta_keys.values())
@@ -235,20 +238,22 @@ class MetaTable:
 
         sql = sql.where(self.sql_table.c.delete_ts.is_(None))
 
-        res_df: DataDF = pd.read_sql_query(
-            sql,
-            con=self.dbconn.con,
-        )
+        with self.dbconn.con.begin() as con:
+            res_df: DataDF = pd.read_sql_query(
+                sql,
+                con=con,
+            )
 
         return data_to_index(res_df, self.primary_keys)
 
     def get_table_debug_info(self) -> TableDebugInfo:
-        return TableDebugInfo(
-            name=self.name,
-            size=self.dbconn.con.execute(
-                select([func.count()]).select_from(self.sql_table)
-            ).fetchone()[0],
-        )
+        with self.dbconn.con.begin() as con:
+            return TableDebugInfo(
+                name=self.name,
+                size=con.execute(
+                    select([func.count()]).select_from(self.sql_table)
+                ).fetchone()[0],
+            )
 
     # TODO Может быть переделать работу с метадатой на контекстный менеджер?
     # FIXME поправить возвращаемые структуры данных, _meta_df должны содержать только _meta колонки
@@ -324,16 +329,17 @@ class MetaTable:
         if len(df) > 0:
             logger.debug(f"Inserting {len(df)} rows into {self.name} data")
 
-            df.to_sql(
-                name=self.sql_table.name,
-                con=self.dbconn.con,
-                schema=self.dbconn.schema,
-                if_exists="append",
-                index=False,
-                chunksize=1000,
-                method="multi",
-                dtype=sql_schema_to_sqltype(self.sql_schema),
-            )
+            with self.dbconn.con.begin() as con:
+                df.to_sql(
+                    name=self.sql_table.name,
+                    con=con,
+                    schema=self.dbconn.schema,
+                    if_exists="append",
+                    index=False,
+                    chunksize=1000,
+                    method="multi",
+                    dtype=sql_schema_to_sqltype(self.sql_schema),
+                )
 
     def _delete_rows(self, df: MetadataDF) -> None:
         if len(df) == 0:
@@ -363,7 +369,8 @@ class MetaTable:
                     )
                 )
 
-            self.dbconn.con.execute(chunk_sql)
+            with self.dbconn.con.begin() as con:
+                con.execute(chunk_sql)
 
     def _update_existing_metadata_rows(self, df: MetadataDF) -> None:
         if len(df) == 0:
@@ -401,6 +408,7 @@ class MetaTable:
                 values_params.append(f'({", ".join(row_values)})')
                 params.update(row_params)
 
+            # TODO сделать через sqlalchemy
             stmt = text(
                 f"""
                 UPDATE {table}
@@ -412,7 +420,8 @@ class MetaTable:
             """
             )
 
-            self.dbconn.con.execution_options(compiled_cache=None).execute(stmt, params)
+            with self.dbconn.con.execution_options(compiled_cache=None).begin() as con:
+                con.execute(stmt, params)
 
     # TODO объединить
     def insert_meta_for_store_chunk(self, new_meta_df: MetadataDF) -> None:
@@ -462,15 +471,17 @@ class MetaTable:
             sql, self.sql_table, self.primary_keys, run_config
         )
 
-        return cast(
-            Iterator[IndexDF],
-            pd.read_sql_query(sql, con=self.dbconn.con, chunksize=1000),
-        )
+        with self.dbconn.con.begin() as con:
+            return cast(
+                Iterator[IndexDF],
+                list(pd.read_sql_query(sql, con=con, chunksize=1000)),
+            )
 
 
 TRANSFORM_META_SCHEMA = [
     Column("process_ts", Float),  # Время последней успешной обработки
     Column("is_success", Boolean),  # Успешно ли обработана строка
+    Column("priority", Integer),  # Приоритет обработки (чем больше, тем выше)
     Column("error", String),  # Текст ошибки
 ]
 
@@ -499,6 +510,35 @@ class TransformMetaTable:
         if create_table:
             self.sql_table.create(self.dbconn.con, checkfirst=True)
 
+    def insert_rows(
+        self,
+        idx: IndexDF,
+    ) -> None:
+        """
+        Создает строки в таблице метаданных для указанных индексов. Если строки
+        уже существуют - не делает ничего.
+        """
+
+        idx = cast(IndexDF, idx[self.primary_keys])
+
+        insert_sql = self.dbconn.insert(self.sql_table).values(
+            [
+                {
+                    "process_ts": 0,
+                    "is_success": False,
+                    "priority": 0,
+                    "error": None,
+                    **idx_dict,  # type: ignore
+                }
+                for idx_dict in idx.to_dict(orient="records")
+            ]
+        )
+
+        sql = insert_sql.on_conflict_do_nothing(index_elements=self.primary_keys)
+
+        with self.dbconn.con.begin() as con:
+            con.execute(sql)
+
     def mark_rows_processed_success(
         self,
         idx: IndexDF,
@@ -512,6 +552,7 @@ class TransformMetaTable:
                 {
                     "process_ts": process_ts,
                     "is_success": True,
+                    "priority": 0,
                     "error": None,
                     **idx_dict,  # type: ignore
                 }
@@ -529,7 +570,8 @@ class TransformMetaTable:
         )
 
         # execute
-        self.dbconn.con.execute(sql)
+        with self.dbconn.con.begin() as con:
+            con.execute(sql)
 
     def mark_rows_processed_error(
         self,
@@ -545,6 +587,7 @@ class TransformMetaTable:
                 {
                     "process_ts": process_ts,
                     "is_success": False,
+                    "priority": 0,
                     "error": error,
                     **idx_dict,  # type: ignore
                 }
@@ -562,10 +605,8 @@ class TransformMetaTable:
         )
 
         # execute
-        self.dbconn.con.execute(sql)
-
-    def reset_metadata(self):
-        self.dbconn.con.execute(self.sql_table.update().values(process_ts=0))
+        with self.dbconn.con.begin() as con:
+            con.execute(sql)
 
     def get_metadata_size(self) -> int:
         """
@@ -578,3 +619,27 @@ class TransformMetaTable:
         sql = select([func.count()]).select_from(self.sql_table)
         res = self.dbconn.con.execute(sql).fetchone()
         return res[0]
+
+    def mark_all_rows_unprocessed(
+        self,
+        run_config: Optional[RunConfig] = None,
+    ) -> None:
+        update_sql = (
+            update(self.sql_table)
+            .values(
+                {
+                    "process_ts": 0,
+                    "is_success": False,
+                    "error": None,
+                }
+            )
+            .where(self.sql_table.c.is_success == True)
+        )
+
+        sql = sql_apply_runconfig_filter(
+            update_sql, self.sql_table, self.primary_keys, run_config
+        )
+
+        # execute
+        with self.dbconn.con.begin() as con:
+            con.execute(sql)