+++ examples/datatable_batch_transform/app.py
@@ -2,17 +2,16 @@ from typing import Dict, List, Optional
 
 import numpy as np
 import pandas as pd
-from datapipe_app import DatapipeApp
-from sqlalchemy import Integer
-from sqlalchemy.sql import functions, select
-from sqlalchemy.sql.schema import Column
-
 from datapipe.compute import Catalog, Pipeline, Table
 from datapipe.core_steps import BatchGenerate, DatatableBatchTransform
 from datapipe.datatable import DataStore, DataTable
 from datapipe.run_config import RunConfig
 from datapipe.store.database import DBConn, TableStoreDB
 from datapipe.types import IndexDF
+from datapipe_app import DatapipeApp
+from sqlalchemy import Integer
+from sqlalchemy.sql import functions, select
+from sqlalchemy.sql.schema import Column
 
 dbconn = DBConn("sqlite+pysqlite3:///db.sqlite")
 
@@ -85,10 +84,11 @@ def count_tbl(
         .group_by(tbl.c["group_id"])
     )
 
-    return pd.read_sql_query(
-        sql,
-        con=ds.meta_dbconn.con,
-    )
+    with ds.meta_dbconn.con.begin() as con:
+        return pd.read_sql_query(
+            sql,
+            con=con,
+        )
 
 
 pipeline = Pipeline(